refactor: rename `page-break-inserter.ts` to `PageBreakInserter.ts`
